+++ src/main/java/video/api/java/sdk/domain/analytics/Info.java
@@ -1,14 +1,15 @@
 package video.api.java.sdk.domain.analytics;
 
+import java.util.Calendar;
 import java.util.Map;
 
 public class Info {
     public final String              sessionId;
-    public final String              loadedAt;
-    public final String              endedAt;
+    public final Calendar            loadedAt;
+    public final Calendar            endedAt;
     public final Map<String, String> metadata;
 
-    public Info(String sessionId, String loadedAt, String endedAt, Map<String, String> metadata) {
+    public Info(String sessionId, Calendar loadedAt, Calendar endedAt, Map<String, String> metadata) {
         this.sessionId = sessionId;
         this.loadedAt  = loadedAt;
         this.endedAt   = endedAt;