+++ src/main/java/video/api/java/sdk/domain/analytics/PlayerSessionEvent.java
@@ -1,13 +1,15 @@
 package video.api.java.sdk.domain.analytics;
 
-public class PlayerSessionEvent {
-    public final String  type;
-    public final String  emittedAt;
-    public final Integer at;
-    public final Integer from;
-    public final Integer to;
+import java.util.Calendar;
 
-    public PlayerSessionEvent(String type, String emittedAt, Integer at, Integer from, Integer to) {
+public class PlayerSessionEvent {
+    public final String   type;
+    public final Calendar emittedAt;
+    public final Integer  at;
+    public final Integer  from;
+    public final Integer  to;
+
+    public PlayerSessionEvent(String type, Calendar emittedAt, Integer at, Integer from, Integer to) {
         this.type      = type;
         this.emittedAt = emittedAt;
         this.at        = at;