+++ src/main/java/video/api/java/sdk/domain/video/Video.java
@@ -2,10 +2,7 @@ package video.api.java.sdk.domain.video;
 
 import video.api.java.sdk.domain.asset.Assets;
 
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
+import java.util.*;
 
 public class Video {
     public static class Source {
@@ -19,8 +16,8 @@ public class Video {
     }
 
     public String              videoId;
-    public String              publishedAt;
-    public String              updatedAt;
+    public Calendar            publishedAt;
+    public Calendar            updatedAt;
     public Source              source;
     public Assets              assets    = new Assets();
     public String              playerId;