+++ src/test/java/video/api/java/sdk/infrastructure/unirest/analytics/PlayerSessionJsonSerializerTest.java
@@ -21,122 +21,81 @@ class PlayerSessionJsonSerializerTest {
 
     @Test
     void deserializeMax() {
-
-        JSONObject data = new JSONObject(
-                "        {\n" +
-                        "            \"session\": {\n" +
-                        "                \"sessionId\": \"ps3uBmLoqYV2O4k9BOeSQ6Sm\",\n" +
-                        "                \"loadedAt\": \"2019-09-09 06:41:06.683+00\",\n" +
-                        "                \"endedAt\": \"2019-09-09 06:41:27.606+00\",\n" +
-                        "            },\n" +
-                        "            \"location\": {\n" +
-                        "                \"country\": \"France\",\n" +
-                        "                \"city\": \"Vaulx-en-Velin\"\n" +
-                        "            },\n" +
-                        "            \"referrer\": {\n" +
-                        "                \"url\": \"unknown\",\n" +
-                        "                \"medium\": \"unknown\",\n" +
-                        "                \"source\": \"unknown\",\n" +
-                        "                \"searchTerm\": \"unknown\"\n" +
-                        "            },\n" +
-                        "            \"device\": {\n" +
-                        "                \"type\": \"smartphone\",\n" +
-                        "                \"vendor\": \"Huawei\",\n" +
-                        "                \"model\": \"P8 Lite (2017)\"\n" +
-                        "            },\n" +
-                        "            \"os\": {\n" +
-                        "                \"name\": \"unknown\",\n" +
-                        "                \"shortname\": \"unknown\",\n" +
-                        "                \"version\": \"unknown\"\n" +
-                        "            },\n" +
-                        "            \"client\": {\n" +
-                        "                \"type\": \"browser\",\n" +
-                        "                \"name\": \"Chrome Mobile\",\n" +
-                        "                \"version\": \"62.0\"\n" +
-                        "            }\n" +
-                        "        }"
-        );
-
-        data.getJSONObject("session").put("metadata", new JSONArray(new ArrayList<JSONObject>() {{
-            add(new JSONObject().put("key", "age").put("value", (String) null));
-            add(new JSONObject().put("key", "name").put("value", "Foo"));
-        }}));
-
-        PlayerSession playerSession = playerSessionJsonSerializer.deserialize(data);
+        PlayerSession playerSession = playerSessionJsonSerializer.deserialize(createSerializedPlayerSessionMax());
         assertEquals(null, playerSession.info.metadata.get("age"));
         assertEquals("Foo", playerSession.info.metadata.get("name"));
+        assertEquals(1574085943808L, playerSession.info.endedAt.getTimeInMillis());
         //assertEquals("viSuccess", playerSession.);
     }
 
     @Test
     void deserializeMin() {
+        PlayerSession playerSession = playerSessionJsonSerializer.deserialize(createSerializedPlayerSessionMin());
 
-        JSONObject analytic = new JSONObject("{\n" +
-                                                     "    \"video\": {\n" +
-                                                     "        \"videoId\": \"viSuccess\",\n" +
-                                                     "        \"title\": \"updated video\",\n" +
-                                                     "    },\n" +
-                                                     "    \"period\": \"2019-09-09\",\n" +
-                                                     "    \"data\": [\n" +
-                                                     "        {\n" +
-                                                     "            \"session\": {\n" +
-                                                     "                \"sessionId\": \"ps3uBmLoqYV2O4k9BOeSQ6Sm\",\n" +
-                                                     "                \"loadedAt\": \"2019-09-09 06:41:06.683+00\",\n" +
-                                                     "                \"endedAt\": \"2019-09-09 06:41:27.606+00\",\n" +
-                                                     "                \"metadata\": [\n" +
-                                                     "                    {\n" +
-                                                     "                        \"key\": \"age\",\n" +
-                                                     "                        \"value\": null\n" +
-                                                     "                    }\n" +
-                                                     "                ]\n" +
-                                                     "            },\n" +
-                                                     "            \"device\": {\n" +
-                                                     "                \"type\": \"smartphone\",\n" +
-                                                     "                \"vendor\": \"Huawei\",\n" +
-                                                     "                \"model\": \"P8 Lite (2017)\"\n" +
-                                                     "            },\n" +
-                                                     "        }\n" +
-                                                     "    ]\n" +
-                                                     "}");
-
-        PlayerSession playerSession = playerSessionJsonSerializer.deserialize(analytic);
-        //assertEquals("viSuccess", playerSession.videoId);
-
+        assertEquals("ps3uBmLoqYV2O4k9BOeSQ6Sm", playerSession.info.sessionId);
 
     }
 
 
     @Test
-    void deserializeAll() {
+    void deserializeCollection() {
         JSONArray collection = new JSONArray();
 
-        JSONObject item = new JSONObject(
-                "{\n" +
-                        "            \"session\": {\n" +
-                        "                \"sessionId\": \"ps3uBmLoqYV2O4k9BOeSQ6Sm\",\n" +
-                        "                \"loadedAt\": \"2019-09-09 06:41:06.683+00\",\n" +
-                        "                \"endedAt\": \"2019-09-09 06:41:27.606+00\",\n" +
-                        "                \"metadata\": [\n" +
-                        "                    {\n" +
-                        "                        \"key\": \"age\",\n" +
-                        "                        \"value\": null\n" +
-                        "                    }\n" +
-                        "                ]\n" +
-                        "            },\n" +
-                        "            \"device\": {\n" +
-                        "                \"type\": \"smartphone\",\n" +
-                        "                \"vendor\": \"Huawei\",\n" +
-                        "                \"model\": \"P8 Lite (2017)\"\n" +
-                        "            },\n" +
-                        "        }\n"
-        );
-
-        collection.put(item);
-        collection.put(item);
-        collection.put(item);
-        collection.put(item);
+        collection.put(createSerializedPlayerSessionMin());
+        collection.put(createSerializedPlayerSessionMin());
+        collection.put(createSerializedPlayerSessionMin());
+        collection.put(createSerializedPlayerSessionMin());
         assertEquals(4, playerSessionJsonSerializer.deserialize(collection).size());
 
+
+    }
+
+    private JSONObject createSerializedPlayerSessionMin() {
+        return new JSONObject()
+                .put("session", new JSONObject()
+                        .put("sessionId", "ps3uBmLoqYV2O4k9BOeSQ6Sm")
+                        .put("loadedAt", "2019-11-18T14:05:43.808000+00:00")
+                        .put("endedAt", "2019-11-18T14:06:20.342000+00:00")
+                );
+    }
+
+    private JSONObject createSerializedPlayerSessionMax() {
+        return new JSONObject()
+                .put("session", new JSONObject()
+                        .put("sessionId", "ps3uBmLoqYV2O4k9BOeSQ6Sm")
+                        .put("loadedAt", "2019-11-18T14:05:43.808000+00:00")
+                        .put("endedAt", "2019-11-18T14:06:20.342000+00:00")
+                        .put("metadata", new JSONArray()
+                                .put(new JSONObject().put("key", "age").put("value", (String) null))
+                                .put(new JSONObject().put("key", "name").put("value", "Foo"))
+                        )
+                )
+                .put("location", new JSONObject()
+                        .put("country", "France")
+                        .put("city", "Vaulx-en-Velin")
+                )
+                .put("referrer", new JSONObject()
+                        .put("url", "unknown")
+                        .put("medium", "unknown")
+                        .put("source", "unknown")
+                        .put("searchTerm", "unknown")
+                )
+                .put("device", new JSONObject()
+                        .put("type", "smartphone")
+                        .put("vendor", "Huawei")
+                        .put("model", "P8 Lite (2017)")
+                )
+                .put("os", new JSONObject()
+                        .put("name", "unknown")
+                        .put("shortname", "unknown")
+                        .put("version", "unknown")
+                )
+                .put("client", new JSONObject()
+                        .put("type", "browser")
+                        .put("name", "Chrome Mobile")
+                        .put("version", "62.0")
+                )
+                ;
 
     }
 }